Database changes

+++ src/main/java/com/example/demo/repopsitory/TransactionRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TransactionRepo extends JpaRepository<Transaction,Long>{
+public interface TransactionRepo extends JpaRepository<Transaction,Integer>{
 
 }
